Separated arena and maq.
